Crate routing

P2PNode implementation for a resilient decentralised network infrastructure.

This is the "engine room" of a hybrid p2p network, where the p2p nodes are built on top of this library. The features this library gives us is:

Sybil resistant p2p nodes

Sharded network with up to approx 200 p2p nodes per shard

All data encrypted at network level with TLS 1.3

Network level quic compatibility, satisfying industry standards and further obfuscating the p2p network data.

Upgrade capable nodes.

All network messages signed via ED25519 and/or BLS

Section consensus via an ABFT algorithm (PARSEC)

Modules

event
Routing events.

Macros

debug
Log a message at the debug level prefixed with the current log ident.

error
Log a message at the error level prefixed with the current log ident.

info
Log a message at the info level prefixed with the current log ident.

log
Log a message at the given level prefixed with the current log ident.

log_or_panic
This macro will panic with the given message if compiled with "mock_base", otherwise it will simply log the message at the requested level.

trace
Log a message at the trace level prefixed with the current log ident.

warn
Log a message at the warn level prefixed with the current log ident.

Structs

FullId
Network identity component containing name, and public and private keys.

Node
Interface for sending and receiving messages to and from other nodes, in the role of a full routing node.

NodeConfig
Node configuration.

P2pNode
Network p2p node identity. When a node knows another node as a P2pNode it's implicitly connected to it. This is separate from being connected at the network layer, which currently is handled by quic-p2p.

PausedState
A type that wraps the internal state of a node while it is paused in order to be upgraded and/or restarted. A value of this type is obtained by pausing a node and can be then used to resume it.

Prefix
A section prefix, i.e. a sequence of bits specifying the part of the network's name space consisting of all names that start with this sequence.

PublicId
Network identity component containing name and public keys.

TransportConfig
QuicP2p configurations

XorName
A XOR_NAME_BITS-bit number, viewed as a point in XOR space.

Enums

DstLocation
Message destination location.

RoutingError
Internal error.

SrcLocation
Message source location.

TransportEvent
QuicP2p Events to the user

Constants

XOR_NAME_LEN
Constant byte length of XorName.