Enum routing::client_errors::MutationError

pub enum MutationError {
    NoSuchAccount,
    AccountExists,
    NoSuchData,
    DataExists,
    DataTooLarge,
    LowBalance,
    InvalidSuccessor,
    InvalidOperation,
    NetworkFull,
    NetworkOther(String),
}

Errors in Put/Post/Delete (mutating) operations involving Core and Vaults

Variants
NoSuchAccount

SAFE Account does not exist for client

AccountExists

Attempt to take an account network name that already exists

NoSuchData

Requested data not found

DataExists

Attempt to create a mutable data when data with such a name already exists

DataTooLarge

Attempt to create/post a data exceeds size limit

LowBalance

Insufficient balance for performing a given mutating operation

InvalidSuccessor

Invalid successor for performing a given mutating operation, e.g. signature mismatch or invalid data versioning

InvalidOperation

Invalid Operation such as a POST on ImmutableData

NetworkFull

The loss of sacrificial copies indicates the network as a whole is no longer having enough space to accept further put request so have to wait for more nodes to join

NetworkOther(String)

Network error occurring at Vault level which has no bearing on clients, e.g. serialisation failure or database failure