Enum routing::Authority

pub enum Authority {
    ClientManager(XorName),
    NaeManager(XorName),
    NodeManager(XorName),
    ManagedNode(XorName),
    Client {
        client_key: PublicKey,
        peer_id: PeerId,
        proxy_node_name: XorName,
    },
}

An entity that can act as a source or destination of a message.

An Authority can be an individual Client or ManagedNode, or a group of nodes, like a NodeManager, ClientManager or NaeManager.

Variants
ClientManager(XorName)

Manager of a Client. XorName is the hash of the Client's client_key.

NaeManager(XorName)

Manager of a network-addressable element, i.e. the group matching this name. XorName is the name of the element in question.

NodeManager(XorName)

Manager of a ManagedNode. XorName is that of the ManagedNode.

ManagedNode(XorName)

A non-client node (i.e. a vault) which is managed by NodeManagers. XorName is provided by the network relocation process immediately after bootstrapping.

Client

A Client.

Fields

client_key: PublicKey

The client's public signing key. The hash of this specifies the location of the Client in the network address space.

peer_id: PeerId

The Crust peer ID of the client.

proxy_node_name: XorName

The name of the single ManagedNode which the Client connects to and proxies all messages through.

Methods

impl Authority

fn is_group(&self) -> bool

Returns true if group authority, otherwise false.

fn is_client(&self) -> bool

Returns true if a client, false if a node or group.

fn name(&self) -> &XorName

Returns the name of authority.

fn to_destination(&self) -> Destination<XorName>

Returns the Destination for the RoutingTable.