Enum routing::Event

pub enum Event {
    Request(RequestMessage),
    Response(ResponseMessage),
    Refresh(u64, Authority, Vec<Vec<u8>>),
    Churn(Vec<XorName>),
    DoRefresh(u64, Authority, XorName),
    Connected,
    Disconnected,
    Terminated,
}

An Event is received at the effective close group of B of a message flow < A | B >

Variants

Request

Request.

Response

Response.

Refresh

Refresh reports to the user the collected accounts for a given refresh event, the arguments are type_tag:u64, authority: Authority, vector_of_bytes: Vec>

Churn

Churn reports whenever our close group is changed, and provides our new close group as a Vec and the name of the node that joined or left our close group as XorName. Our close group is sorted from our name and always includes our own name as the first element.

DoRefresh

DoRefresh reports that a Refresh message is circulating the effective close group of the given Authority, but that the user is outside of the close group of the churn that initiated the call for refresh. To ensure that the account of the current user is also accumulated a DoRefresh indicates precisely one account routing will expect the user to do a ::routing::request_refresh for, if a matching account is held by the user.

Connected

Connected.

Disconnected

Disconnected.

Terminated

Event::Terminated is called after RoutingNode::stop() has terminated internal processes