Enum routing::InterfaceError

pub enum InterfaceError {
    NotConnected,
    ChannelRxError(RecvError),
    EventSenderError(EventSenderError<MaidSafeEventCategory, Action>),
}

InterfaceError.

Variants

NotConnected

NotConnected.

ChannelRxError

Error while trying to receive a message from a channel

EventSenderError

Error while trying to transmit an event via a channel