Enum routing::RequestContent

pub enum RequestContent {
    GetNetworkName {
        current_id: PublicId,
    },
    ExpectCloseNode {
        expect_id: PublicId,
    },
    GetCloseGroup,
    Connect,
    Endpoints {
        encrypted_endpoints: Vec<u8>,
        nonce_bytes: [u8; box_::NONCEBYTES],
    },
    GetPublicId,
    GetPublicIdWithEndpoints {
        encrypted_endpoints: Vec<u8>,
        nonce_bytes: [u8; box_::NONCEBYTES],
    },
    Refresh(Vec<u8>),
    Get(DataRequest, MessageId),
    Put(Data, MessageId),
    Post(Data, MessageId),
    Delete(Data, MessageId),
}

The request types

Variants

GetNetworkName

Ask the network to alter your PublicId name.

This is sent by a Client to its NaeManager with the intent to become a routing node with a new name chosen by the NaeManager.

Fields

current_id

The client's PublicId (public keys and name)

ExpectCloseNode

Notify a joining node's NodeManager so that it expects a GetCloseGroup request from it.

Fields

expect_id

The joining node's PublicId (public keys and name)

GetCloseGroup

Request the PublicIds of the recipient's close group.

This is sent from a joining node to its NodeManager to request the PublicIds of the NodeManager's members.

Connect

Request a direct connection to the recipient.

Endpoints

Send our endpoints encrypted to a node we wish to connect to and have the keys for.

Fields

encrypted_endpoints

Encrypted crust endpoints (socket address and protocol).

nonce_bytes

Nonce used to provide a salt in the encrytped message.

GetPublicId

Ask each member of a group near a node address for the PublicId.

GetPublicIdWithEndpoints

Ask for a PublicId but provide our endpoints encrytped.

Fields

encrypted_endpoints

Encrypted crust endpoints (socket address and protocol).

nonce_bytes

Nonce used to provide a salt in the encrytped message.

Refresh

Message from upper layers sending network state on any network churn event.

Get

Ask for data from network, passed from API with data name as parameter

Put

Put data to network. Provide actual data as parameter

Post

Post data to network. Provide actual data as parameter

Delete

Delete data from network. Provide actual data as parameter