Enum routing::Response

pub enum Response {
    GetSuccess(Data, MessageId),
    PutSuccess(DataIdentifier, MessageId),
    PostSuccess(DataIdentifier, MessageId),
    DeleteSuccess(DataIdentifier, MessageId),
    AppendSuccess(DataIdentifier, MessageId),
    GetAccountInfoSuccess {
        id: MessageId,
        data_stored: u64,
        space_available: u64,
    },
    GetFailure {
        id: MessageId,
        data_id: DataIdentifier,
        external_error_indicator: Vec<u8>,
    },
    PutFailure {
        id: MessageId,
        data_id: DataIdentifier,
        external_error_indicator: Vec<u8>,
    },
    PostFailure {
        id: MessageId,
        data_id: DataIdentifier,
        external_error_indicator: Vec<u8>,
    },
    DeleteFailure {
        id: MessageId,
        data_id: DataIdentifier,
        external_error_indicator: Vec<u8>,
    },
    AppendFailure {
        id: MessageId,
        data_id: DataIdentifier,
        external_error_indicator: Vec<u8>,
    },
    GetAccountInfoFailure {
        id: MessageId,
        external_error_indicator: Vec<u8>,
    },
}

Response message types

Variants
GetSuccess(Data, MessageId)

Reply with the requested data (may not be ignored)

Sent from a ManagedNode to an NaeManager, and from there to a Client, although this may be shortcut if the data is in a node's cache.

PutSuccess(DataIdentifier, MessageId)

Success token for Put (may be ignored)

PostSuccess(DataIdentifier, MessageId)

Success token for Post (may be ignored)

DeleteSuccess(DataIdentifier, MessageId)

Success token for delete (may be ignored)

AppendSuccess(DataIdentifier, MessageId)

Success token for append (may be ignored)

GetAccountInfoSuccess

Response containing account information for requested Client account

Fields

id: MessageId

Unique message identifier

data_stored: u64

Amount of data stored on the network by this Client

space_available: u64

Amount of network space available to this Client

GetFailure

Error for Get, includes signed request to prevent injection attacks

Fields

id: MessageId

Unique message identifier

data_id: DataIdentifier

ID of the affected data chunk

external_error_indicator: Vec<u8>

Error type sent back, may be injected from upper layers

PutFailure

Error for Put, includes signed request to prevent injection attacks

Fields

id: MessageId

Unique message identifier

data_id: DataIdentifier

ID of the affected data chunk

external_error_indicator: Vec<u8>

Error type sent back, may be injected from upper layers

PostFailure

Error for Post, includes signed request to prevent injection attacks

Fields

id: MessageId

Unique message identifier

data_id: DataIdentifier

ID of the affected data chunk

external_error_indicator: Vec<u8>

Error type sent back, may be injected from upper layers

DeleteFailure

Error for delete, includes signed request to prevent injection attacks

Fields

id: MessageId

Unique message identifier

data_id: DataIdentifier

ID of the affected data chunk

external_error_indicator: Vec<u8>

Error type sent back, may be injected from upper layers

AppendFailure

Error for append, includes signed request to prevent injection attacks

Fields

id: MessageId

Unique message identifier

data_id: DataIdentifier

ID of the affected data chunk

external_error_indicator: Vec<u8>

Error type sent back, may be injected from upper layers

GetAccountInfoFailure

Error for GetAccountInfo

Fields

id: MessageId

Unique message identifier

external_error_indicator: Vec<u8>

Error type sent back, may be injected from upper layers

Methods

impl Response

fn priority(&self) -> u8

The priority Crust should send this message with.

fn is_cacheable(&self) -> bool

Is this response cacheable?