Enum routing::ResponseContent

pub enum ResponseContent {
    GetNetworkName {
        relocated_id: PublicId,
    },
    GetPublicId {
        public_id: PublicId,
    },
    GetPublicIdWithEndpoints {
        public_id: PublicId,
        encrypted_endpoints: Vec<u8>,
        nonce_bytes: [u8; box_::NONCEBYTES],
    },
    GetCloseGroup {
        close_group_ids: Vec<PublicId>,
    },
    GetSuccess(Data, MessageId),
    PutSuccess(Digest, MessageId),
    PostSuccess(Digest, MessageId),
    DeleteSuccess(Digest, MessageId),
    GetFailure {
        id: MessageId,
        request: RequestMessage,
        external_error_indicator: Vec<u8>,
    },
    PutFailure {
        id: MessageId,
        request: RequestMessage,
        external_error_indicator: Vec<u8>,
    },
    PostFailure {
        id: MessageId,
        request: RequestMessage,
        external_error_indicator: Vec<u8>,
    },
    DeleteFailure {
        id: MessageId,
        request: RequestMessage,
        external_error_indicator: Vec<u8>,
    },
}

The response types

All responses map to a specific request, and where the request was from a single node or client, the response will contain the signed request to prevent forgery.

Variants

GetNetworkName

Reply with the new PublicId for the joining node.

Sent from the NaeManager to the Client.

Fields

relocated_id

Supplied PublicId, but with the new name

GetPublicId

Reply with the requested PublicId.

Sent from the NodeManager to the joining node.

Fields

public_id

The requested PublicId

GetPublicIdWithEndpoints

Reply with the PublicId along with the sender's encrypted endpoints

Sent from a ManagedNode to another node or client.

Fields

public_id

Our PublicId

encrypted_endpoints

Their endpoints

nonce_bytes

Message salt

GetCloseGroup

Return the close PublicIds back to the requestor.

Sent from a NodeManager to a node or client.

Fields

close_group_ids

Our close group PublicIds.

GetSuccess

Reply with the requested data (may not be ignored)

Sent from a ManagedNode to an NaeManager, and from there to a Client, although this may be shortcut if the data is in a node's cache.

PutSuccess

Success token for Put (may be ignored)

PostSuccess

Success token for Post (may be ignored)

DeleteSuccess

Success token for delete (may be ignored)

GetFailure

Error for Get, includes signed request to prevent injection attacks

Fields

id

Unique message identifier

request

Originator's signed request

external_error_indicator

Error type sent back, may be injected from upper layers

PutFailure

Error for Put, includes signed request to prevent injection attacks

Fields

id

Unique message identifier

request

Originator's signed request

external_error_indicator

Error type sent back, may be injected from upper layers

PostFailure

Error for Post, includes signed request to prevent injection attacks

Fields

id

Unique message identifier

request

Originator's signed request

external_error_indicator

Error type sent back, may be injected from upper layers

DeleteFailure

Error for delete, includes signed request to prevent injection attacks

Fields

id

Unique message identifier

request

Originator's signed request

external_error_indicator

Error type sent back, may be injected from upper layers