Enum routing::TransportEvent

pub enum TransportEvent {
    BootstrapFailure,
    BootstrappedTo {
        node: SocketAddr,
    },
    ConnectionFailure {
        peer: Peer,
        err: QuicP2pError,
    },
    SentUserMessage {
        peer: Peer,
        msg: Bytes,
        token: u64,
    },
    UnsentUserMessage {
        peer: Peer,
        msg: Bytes,
        token: u64,
    },
    ConnectedTo {
        peer: Peer,
    },
    NewMessage {
        peer: Peer,
        msg: Bytes,
    },
    Finish,
}

QuicP2p Events to the user

Variants

BootstrapFailure

Network bootstrap failed.

BootstrappedTo

Bootstrap connection to this node was successful.

Fields of BootstrappedTo

node: SocketAddr

Node information.

ConnectionFailure

Connection to this peer failed.

Fields of ConnectionFailure

peer: Peer

Peer.

err: QuicP2pError

Error explaining connection failure.

SentUserMessage

The given message was successfully sent to this peer.

Fields of SentUserMessage

peer: Peer

Peer.

msg: Bytes

Sent message.

token: u64

Token, originally given by the user, for context.

UnsentUserMessage

The given message was not sent to this peer.

Fields of UnsentUserMessage

peer: Peer

Peer.

msg: Bytes

Unsent message.

token: u64

Token, originally given by the user, for context.

ConnectedTo

Successfully connected to this peer.

Fields of ConnectedTo

peer: Peer

Peer information.

NewMessage

A new message was received from this peer.

Fields of NewMessage

peer: Peer

Sending peer.

msg: Bytes

The new message.

Finish

No more messages will be fired after this