Enum routing::error::ResponseError

pub enum ResponseError {
    Abort,
    LowBalance(Data, u32),
    InvalidRequest(Data),
    FailedRequestForData(Data),
    HadToClearSacrificial(NameType, u32),
}

Represents response errors.

Variants

Abort

Abort is for user to indicate that the state can be dropped; if received by routing, it will drop the state.

LowBalance

On low balance or no account registered

InvalidRequest

invalid request

FailedRequestForData

failure to complete request for data

HadToClearSacrificial

had to clear Sacrificial Data in order to complete request