Enum routing::event::Event

pub enum Event {
    Request {
        request: ExternalRequest,
        our_authority: Authority,
        from_authority: Authority,
        response_token: Option<SignedToken>,
    },
    Response {
        response: ExternalResponse,
        our_authority: Authority,
        from_authority: Authority,
    },
    FailedRequest {
        request: ExternalRequest,
        our_authority: Option<Authority>,
        location: Authority,
        interface_error: InterfaceError,
    },
    FailedResponse {
        response: ExternalResponse,
        our_authority: Option<Authority>,
        location: Authority,
        interface_error: InterfaceError,
    },
    Refresh(u64, Authority, Vec<Vec<u8>>),
    Churn(Vec<NameType>, NameType),
    DoRefresh(u64, Authority, NameType),
    Bootstrapped,
    Connected,
    Disconnected,
    Terminated,
}

An Event is received at the effective close group of B of a message flow < A | B >

Variants

Request

Request.

Fields

request

External request.

our_authority

Our authority.

from_authority

From authority.

response_token

Not set when the request came from a group.

Response

Response.

Fields

response

External response.

our_authority

Our authority.

from_authority

From authority.

FailedRequest

FailedRequest.

Fields

request

External request.

our_authority

Our authority.

location

From authority.

interface_error

Interface error.

FailedResponse

FailedResponse.

Fields

response

External response.

our_authority

Our authority.

location

From authority.

interface_error

Interface error.

Refresh

Refresh reports to the user the collected accounts for a given refresh event, the arguments are type_tag:u64, authority: Authority, vector_of_bytes: Vec>

Churn

Churn reports whenever our close group is changed, and provides our new close group as a Vec and the name of the node that joined or left our close group as NameType. Our close group is sorted from our name and always includes our own name as the first element.

DoRefresh

DoRefresh reports that a Refresh message is circulating the effective close group of the given Authority, but that the user is outside of the close group of the churn that initiated the call for refresh. To ensure that the account of the current user is also accumulated a DoRefresh indicates precisely one account routing will expect the user to do a ::routing::request_refresh for, if a matching account is held by the user.

Bootstrapped

Bootstrapped.

Connected

Connected.

Disconnected

Disconnected.

Terminated

Event::Terminated is called after RoutingNode::stop() has terminated internal processes