Struct routing::id::Id

pub struct Id {
    // some fields omitted
}

Id.

Methods

impl Id

fn new() -> Id

Contruct new Id.

fn signing_public_key(&self) -> PublicKey

Public signing key.

fn signing_private_key(&self) -> &SecretKey

Secret signing key.

fn encrypting_public_key(&self) -> PublicKey

Public encryption key.

fn with_keys(sign_keys: (PublicKey, SecretKey), encrypt_keys: (PublicKey, SecretKey)) -> Id

Construct with given keys, (Client requirement).

fn name(&self) -> NameType

Original/relocated name.

fn assign_relocated_name(&mut self, relocated_name: NameType) -> bool

Name field is initially same as original_name, this should be later overwritten by relocated name provided by the network using this method

fn is_relocated(&self) -> bool

Checks if the name is updated to a relocated name.