Struct routing::messages::HopMessage

pub struct HopMessage {
    // some fields omitted
}

A wrapper for all messages sent form node to node. Allows nodes to be sure there has been no alteration of message in transit. Also defeats MiTM attacks

Methods

impl HopMessage

fn new(content: SignedMessage, name: XorName, sign_key: &SecretKey) -> Result<HopMessage, RoutingError>

Wrap a signed message for transmission to next hop.

fn verify(&self, verification_key: &PublicKey) -> Result<(), RoutingError>

Validate message is signed by public key contained in message. this does not validate the message came from know node. That requires a check against the routing table of the node to identify the name associated with the PublicKey

fn extract(&self) -> (SignedMessage, XorName)

return a signed message, does not validate message. [#verify] must be called prior to ensure sender is valid and validly signed he message

fn name(&self) -> &XorName

The name asociated with te hop message.