Enum routing::node_interface::MethodCall

pub enum MethodCall {
    None,
    Put {
        destination: NameType,
        content: Box<Sendable>,
    },
    Get {
        type_id: u64,
        name: NameType,
    },
    Post,
    Refresh {
        type_tag: u64,
        from_group: NameType,
        payload: Vec<u8>,
    },
    SendOn {
        destination: NameType,
    },
}

MethodCall denotes a specific request to be carried out by routing.

Variants

None

request for no action

Put

request to have destination to handle put for the content

Fields

destination

content

Get

request to retreive data with specified type and name from network

Fields

type_id

name

Post

request to post

Refresh

request to refresh

Fields

type_tag

from_group

payload

SendOn

request to send on the request to destination for further handling

Fields

destination