Trait routing::node_interface::Interface

pub trait Interface: Sync + Send {
    fn handle_get(&mut self, type_id: u64, name: NameType, our_authority: Authority, from_authority: Authority, from_address: NameType) -> Result<MessageAction, InterfaceError>;
    fn handle_put(&mut self, our_authority: Authority, from_authority: Authority, from_address: NameType, dest_address: DestinationAddress, data: Vec<u8>) -> Result<MessageAction, InterfaceError>;
    fn handle_post(&mut self, our_authority: Authority, from_authority: Authority, from_address: NameType, name: NameType, data: Vec<u8>) -> Result<MessageAction, InterfaceError>;
    fn handle_refresh(&mut self, type_tag: u64, from_group: NameType, payloads: Vec<Vec<u8>>);
    fn handle_get_response(&mut self, from_address: NameType, response: Result<Vec<u8>, ResponseError>) -> MethodCall;
    fn handle_put_response(&mut self, from_authority: Authority, from_address: NameType, response: Result<Vec<u8>, ResponseError>) -> MethodCall;
    fn handle_post_response(&mut self, from_authority: Authority, from_address: NameType, response: Result<Vec<u8>, ResponseError>);
    fn handle_churn(&mut self, close_group: Vec<NameType>) -> Vec<MethodCall>;
    fn handle_cache_get(&mut self, type_id: u64, name: NameType, from_authority: Authority, from_address: NameType) -> Result<MessageAction, InterfaceError>;
    fn handle_cache_put(&mut self, from_authority: Authority, from_address: NameType, data: Vec<u8>) -> Result<MessageAction, InterfaceError>;
}

The Interface trait introduces the methods expected to be implemented by the user of RoutingNode

Required Methods

fn handle_get(&mut self, type_id: u64, name: NameType, our_authority: Authority, from_authority: Authority, from_address: NameType) -> Result<MessageAction, InterfaceError>

depending on our_authority and from_authority, data or address of the node potentially storing data with specified name and type_id is returned, on success. failure to provide data or an address is indicated as an InterfaceError.

fn handle_put(&mut self, our_authority: Authority, from_authority: Authority, from_address: NameType, dest_address: DestinationAddress, data: Vec<u8>) -> Result<MessageAction, InterfaceError>

depending on our_authority and from_authority, data is stored on current node or an address (with different authority) for further handling of the request is provided. failure is indicated as an InterfaceError.

fn handle_post(&mut self, our_authority: Authority, from_authority: Authority, from_address: NameType, name: NameType, data: Vec<u8>) -> Result<MessageAction, InterfaceError>

depending on our_authority and from_authority, post request is handled by current node or an address for further handling of the request is provided. Failure is indicated as an InterfaceError.

fn handle_refresh(&mut self, type_tag: u64, from_group: NameType, payloads: Vec<Vec<u8>>)

Handle messages internal to the group (triggered by churn events). Payloads from these messages are grouped by (type_tag, from_group) key, and once there is enough of them, they are returned in the payloads argument.

fn handle_get_response(&mut self, from_address: NameType, response: Result<Vec<u8>, ResponseError>) -> MethodCall

handles the response to a put request. Depending on ResponseError, performing an action of type MethodCall is requested.

fn handle_put_response(&mut self, from_authority: Authority, from_address: NameType, response: Result<Vec<u8>, ResponseError>) -> MethodCall

handles the response to a put request. Depending on ResponseError, performing an action of type MethodCall is requested.

fn handle_post_response(&mut self, from_authority: Authority, from_address: NameType, response: Result<Vec<u8>, ResponseError>)

handles the response to a post request. Depending on ResponseError, performing an action of type MethodCall is requested.

fn handle_churn(&mut self, close_group: Vec<NameType>) -> Vec<MethodCall>

handles the actions to be carried out in the event of a churn. The function provides a list of actions (of type MethodCall) to be carried out in order to update relevant nodes.

fn handle_cache_get(&mut self, type_id: u64, name: NameType, from_authority: Authority, from_address: NameType) -> Result<MessageAction, InterfaceError>

attempts to potentially retrieve data from cache.

fn handle_cache_put(&mut self, from_authority: Authority, from_address: NameType, data: Vec<u8>) -> Result<MessageAction, InterfaceError>

attempts to store data in cache. The type of data and/or from_authority indicates if store in cache is required.