Struct routing::peer::Peer

pub struct Peer {
    // some fields omitted
}

Peer enables multiple endpoints per peer in the network. It currently wraps around crust::endpoint, and will be extended to enable multiple endpoints, merging, comparing and other functionality.

Methods

impl Peer

fn new(identity: ConnectionName, endpoint: Endpoint) -> Peer

fn identity(&self) -> &ConnectionName

fn endpoint(&self) -> &Endpoint

fn public_id(&self) -> &Option<PublicId>

fn connected_timestamp(&self) -> &SteadyTime

fn set_public_id(&mut self, public_id: PublicId)