Struct routing::public_id::PublicId

pub struct PublicId {
    // some fields omitted
}

PublicId.

Methods

impl PublicId

fn new(id: &Id) -> PublicId

Construct new PublicId.

fn name(&self) -> NameType

Return initial/relocated name.

fn set_name(&mut self, name: NameType)

Set the name to new value.

fn client_name(&self) -> NameType

Return client name derived from public key.

fn serialised_contents(&self) -> Result<Vec<u8>, RoutingError>

Serialise the content of self.

fn assign_relocated_name(&mut self, relocated_name: NameType)

Name field is initially same as original_name, this should be replaced by relocated name calculated by the nodes close to original_name by using this method

fn signing_public_key(&self) -> PublicKey

Return public signing key.

fn is_relocated(&self) -> bool

Checks if the name is updated to a relocated name.