Struct routing::routing::Routing

pub struct Routing {
    // some fields omitted
}

Routing provides an actionable interface to RoutingNode. On constructing a new Routing object a RoutingNode will also be started. Routing objects are clonable for multithreading, or a Routing object can be cloned with a new set of keys while preserving a single RoutingNode.

Methods

impl Routing

fn new(event_sender: Sender<Event>) -> Routing

Starts a new RoutingIdentity, which will also start a new RoutingNode. The RoutingNode will attempt to achieve full routing node status. The intial Routing object will have newly generated keys

fn get_request(&self, our_authority: Authority, location: Authority, data_request: DataRequest)

Send a Get message with a DataRequest to an Authority, signed with given keys.

fn put_request(&self, our_authority: Authority, location: Authority, data: Data)

Add something to the network

fn post_request(&self, our_authority: Authority, location: Authority, data: Data)

Change something already on the network

fn delete_request(&self, our_authority: Authority, location: Authority, data: Data)

Remove something from the network

fn get_response(&self, our_authority: Authority, location: Authority, data: Data, data_request: DataRequest, signed_token: Option<SignedToken>)

Respond to a get_request (no error can be sent) If we received the request from a group, we'll not get the signed_token.

fn put_response(&self, our_authority: Authority, location: Authority, response_error: ResponseError, signed_token: Option<SignedToken>)

response error to a put request

fn post_response(&self, our_authority: Authority, location: Authority, response_error: ResponseError, signed_token: Option<SignedToken>)

Response error to a post request

fn delete_response(&self, our_authority: Authority, location: Authority, response_error: ResponseError, signed_token: Option<SignedToken>)

response error to a delete respons

fn refresh_request(&self, type_tag: u64, our_authority: Authority, content: Bytes, cause: NameType)

Refresh the content in the close group nodes of group address content::name. This method needs to be called when churn is triggered. all the group members need to call this, otherwise it will not be resolved as a valid content. If the authority provided (our_authority) is not a group, the request for refresh will be dropped.

fn set_cache_options(&self, cache_options: CacheOptions)

Dynamically enable/disable caching for Data types.

fn stop(&mut self)

Signal to RoutingNode that it needs to refuse new messages and handle all outstanding messages. After handling all messages it will send an Event::Terminated to the user.