Struct routing::routing_client::RoutingClient

pub struct RoutingClient<F: Interface> {
    // some fields omitted
}

Methods

impl<F> RoutingClient<F> where F: Interface

fn new(my_interface: Arc<Mutex<F>>, id: Id) -> RoutingClient<F>

fn get(&mut self, type_id: u64, name: NameType) -> Result<MessageId, IoError>

Retrieve something from the network (non mutating) - Direct call

fn put<T>(&mut self, content: T) -> Result<MessageId, IoError> where T: Sendable

Add something to the network, will always go via ClientManager group

fn unauthorised_put(&mut self, destination: NameType, content: Box<Sendable>)

Add content to the network

fn run(&mut self)

fn bootstrap(&mut self, bootstrap_list: Option<Vec<Endpoint>>, beacon_port: Option<u16>) -> Result<(), RoutingError>

Use bootstrap to attempt connecting the client to previously known nodes, or use CRUST self-discovery options.