Module routing::routing_membrane

Routing membrane is a single thread responsible for the in- and outgoing messages. It accepts messages received from CRUST. The membrane evaluates whether a message is to be forwarded, or accepted into the membrane as a request where Sentinel holds it until verified and resolved. Requests resolved by Sentinel, will be handed on to the Interface for actioning. A limited number of messages are deliberatly for Routing and network management purposes. Some network management messages are directly handled without Sentinel resolution. Other network management messages are handled by Routing after Sentinel resolution.

Structs

RoutingMembrane
Routing Membrane