Struct routing::routing_membrane::RoutingMembrane

pub struct RoutingMembrane<F: Interface> {
    // some fields omitted
}

Routing Membrane

Methods

impl<F> RoutingMembrane<F> where F: Interface

fn new(cm: ConnectionManager, event_input: Receiver<Event>, bootstrap_endpoint: Option<Endpoint>, accepting_on: Vec<Endpoint>, relocated_id: Id, personas: F) -> RoutingMembrane<F>

fn get(&mut self, type_id: u64, name: NameType)

Retrieve something from the network (non mutating) - Direct call

fn put(&mut self, destination: NameType, content: Box<Sendable>)

Add something to the network, will always go via ClientManager group

fn unauthorised_put(&mut self, destination: NameType, content: Box<Sendable>)

Add something to the network

fn refresh(&mut self, type_tag: u64, from_group: NameType, content: Bytes)

Refresh the content in the close group nodes of group address content::name. This method needs to be called when churn is triggered. all the group members need to call this, otherwise it will not be resolved as a valid content.

fn run(&mut self)

RoutingMembrane::Run starts the membrane