Struct routing::routing_node::RoutingNode

pub struct RoutingNode<F, G> where F: Interface + 'static, G: CreatePersonas<F> {
    // some fields omitted
}

DHT node

Methods

impl<F, G> RoutingNode<F, G> where F: Interface + 'static, G: CreatePersonas<F>

fn new(genesis: G) -> RoutingNode<F, G>

fn run_zero_membrane(&mut self)

Starts a node without requiring responses from the network. Starts the routing membrane without looking to bootstrap. It will relocate its own address with the hash of twice its name. This allows the network to later reject this zero node when the routing_table is full.

A zero_membrane will not be able to connect to an existing network, and as a special node, it will be rejected by the network later on.

fn bootstrap(&mut self, bootstrap_list: Option<Vec<Endpoint>>, beacon_port: Option<u16>) -> Result<(), RoutingError>

Bootstrap the node to an existing (or zero) node on the network. If a bootstrap list is provided those will be used over the beacon support from CRUST. Spawns a new thread and moves a newly constructed Membrane into this thread. Routing node uses the genesis object to create a new instance of the personas to embed inside the membrane.