Struct routing::routing_table::RoutingTable

pub struct RoutingTable {
    // some fields omitted
}

The RoutingTable class is used to maintain a list of contacts to which we are connected.

Methods

impl RoutingTable

fn new(our_id: DhtId) -> RoutingTable

fn get_bucket_size() -> usize

fn get_parallelism() -> usize

fn get_optimal_size() -> usize

fn get_group_size() -> usize

fn get_quorum_size() -> usize

fn add_node(&mut self, their_info: NodeInfo) -> (bool, Option<NodeInfo>)

Potentially adds a contact to the routing table. If the contact is added, the first return arg is true, otherwise false. If adding the contact caused another contact to be dropped, the dropped one is returned in the second field, otherwise the optional field is empty. The following steps are used to determine whether to add the new contact or not:

1 - if the contact is ourself, or doesn't have a valid public key, or is already in the table, it will not be added 2 - if the routing table is not full (size < OptimalSize()), the contact will be added 3 - if the contact is within our close group, it will be added 4 - if we can find a candidate for removal (a contact in a bucket with more than 'BucketSize()' contacts, which is also not within our close group), and if the new contact will fit in a bucket closer to our own bucket, then we add the new contact.

fn check_node(&self, their_id: &DhtId) -> bool

This is used to see whether to bother retrieving a contact's public key from the PKI with a view to adding the contact to our table. The checking procedure is the same as for 'AddNode' above, except for the lack of a public key to check in step 1.

fn drop_node(&mut self, node_to_drop: &DhtId)

This unconditionally removes the contact from the table.

fn target_nodes(&self, target: DhtId) -> Vec<NodeInfo>

This returns a collection of contacts to which a message should be sent onwards. It will return all of our close group (comprising 'GroupSize' contacts) if the closest one to the target is within our close group. If not, it will return the 'Parallelism()' closest contacts to the target.

fn our_close_group(&self) -> Vec<NodeInfo>

This returns our close group, i.e. the 'GroupSize' contacts closest to our ID (or the entire table if we hold less than 'GroupSize' contacts in total).

fn get_public_key(&self, their_id: DhtId) -> Option<PublicKey>

This returns the public key for the given node if the node is in our table.

fn size(&self) -> usize

fn closer_to_target(lhs: &DhtId, rhs: &DhtId, target: &DhtId) -> bool