Module routing::sendable

Traits

Sendable
This trait is required for any type of message to be passed to routing, refresh / account transfer is optional The name will let routing know its a NaeManager and the owner will allow routing to hash the requesters ID with this name (by hashing the requesters ID) for put and post messages