Trait routing::sendable::Sendable

pub trait Sendable {
    fn name(&self) -> NameType;
    fn type_tag(&self) -> u64;
    fn serialised_contents(&self) -> Vec<u8>;
    fn refresh(&self) -> bool;
    fn merge(&self, responses: Vec<Box<Sendable>>) -> Option<Box<Sendable>>;

    fn owner(&self) -> Option<NameType> { ... }
}

This trait is required for any type of message to be passed to routing, refresh / account transfer is optional The name will let routing know its a NaeManager and the owner will allow routing to hash the requesters ID with this name (by hashing the requesters ID) for put and post messages

Required Methods

fn name(&self) -> NameType

fn type_tag(&self) -> u64

fn serialised_contents(&self) -> Vec<u8>

fn refresh(&self) -> bool

fn merge(&self, responses: Vec<Box<Sendable>>) -> Option<Box<Sendable>>

Provided Methods

fn owner(&self) -> Option<NameType>