Struct routing::Client

pub struct Client {
    // some fields omitted
}

Interface for sending and receiving messages to and from a network of nodes in the role of a client.

A client is connected to the network via one or more nodes. Messages are never routed via a client, and a client cannot be part of a group authority.

Methods

impl Client

fn new(event_sender: Sender<Event>, keys: Option<FullId>) -> Result<Client, RoutingError>

Create a new Client.

It will automatically connect to the network, but not attempt to achieve full routing node status. The name of the client will be the name of the PublicId of the keys and must equal the SHA512 hash of its public signing key, otherwise the client will be instantly terminated.

Keys will be exchanged with the ClientAuthority so that communication with the network is cryptographically secure and uses group consensus. The restriction for the client name exists to ensure that the client cannot choose its ClientAuthority.

fn send_get_request(&mut self, dst: Authority, data_request: DataRequest) -> Result<(), InterfaceError>

Send a Get message with a DataRequest to an Authority, signed with given keys.

fn send_put_request(&self, dst: Authority, data: Data) -> Result<(), InterfaceError>

Add something to the network

fn send_post_request(&self, dst: Authority, data: Data) -> Result<(), InterfaceError>

Change something already on the network

fn send_delete_request(&self, dst: Authority, data: Data) -> Result<(), InterfaceError>

Remove something from the network