Struct routing::Node

pub struct Node { /* fields omitted */ }

Interface for sending and receiving messages to and from other nodes, in the role of a full routing node.

A node is a part of the network that can route messages and be a member of a section or group location. Its methods can be used to send requests and responses as either an individual Node or as a part of a section or group location. Their src argument indicates that role, and can be any SrcLocation.

Methods

impl Node

pub fn new(
config: NodeConfig
) -> (Self, Receiver<Event>, Receiver<TransportEvent>)

Create new node using the given config.

Returns the node itself, the user event receiver and the client network event receiver.

pub fn pause(self) -> Result<PausedState, RoutingError>

Pauses the node in order to be upgraded and/or restarted. Returns InvalidState error if the node is not a member of any section yet.

pub fn resume(state: PausedState) -> (Self, Receiver<Event>)

Resume previously paused node.

pub fn register<'a>(&'a mut self, select: &mut Select<'a>)

Register the node event channels with the provided selector.

pub fn handle_selected_operation(
&mut self,
op_index: usize
) -> Result<(), RecvError>

Processes events received externally from one of the channels. For this function to work properly, the node event channels need to be registered by calling ApprovedPeer::register. Select::ready needs to be called to get op_index, the event channel index.

This function is non-blocking.

Errors are permanent failures due to either: node termination, the permanent closing of one of the event channels, or an invalid (unknown) channel index.

pub fn is_running(&self) -> bool

Returns whether this node is running or has been terminated.

pub fn id(&self) -> &PublicId

Returns the PublicId of this node.

pub fn name(&self) -> &XorName

The name of this node.

pub fn our_connection_info(&mut self) -> Result<SocketAddr, RoutingError>

Returns connection info of this node.

pub fn our_prefix(&self) -> Option<&Prefix<XorName>>

Our Prefix once we are a part of the section.

pub fn matches_our_prefix(&self, name: &XorName) -> Result<bool, RoutingError>

Finds out if the given XorName matches our prefix. Returns error if we don't have a prefix because we haven't joined any section yet.

pub fn is_elder(&self) -> bool

Returns whether the node is Elder.

pub fn our_elders(&self) -> impl Iterator<Item = &P2pNode>

Returns the information of all the current section elders.

pub fn closest_known_elders_to<'a>(
&'a self,
name: &'a XorName
) -> impl Iterator<Item = &'a P2pNode>

Find out the closest Elders to a given XorName that we know of.

Note that the Adults of a section only know about their section Elders. Hence they will always return the section Elders' info.

pub fn our_adults(&self) -> impl Iterator<Item = &P2pNode>

Returns the information of all the current section adults.

pub fn our_adults_sorted_by_distance_to(&self, name: &XorName) -> Vec<&P2pNode>

Returns the adults of our section sorted by their distance to name (closest first). If we are not elder or if there are no adults in the section, returns empty vec.

pub fn in_dst_location(&self, dst: &DstLocation) -> bool

Checks whether the given location represents self.

pub fn vote_for_user_event(
&mut self,
event: Vec<u8>
) -> Result<(), RoutingError>

Vote for a user-defined event. Returns InvalidState error if we are not an elder.

pub fn send_message(
&mut self,
src: SrcLocation,
dst: DstLocation,
content: Vec<u8>
) -> Result<(), RoutingError>

Send a message.

pub fn send_message_to_client(
&mut self,
peer_addr: SocketAddr,
msg: Bytes,
token: Token
) -> Result<(), RoutingError>

Send a message to a client peer.

pub fn disconnect_from_client(
&mut self,
peer_addr: SocketAddr
) -> Result<(), RoutingError>

Disconnect form a client peer.

pub fn public_key_set(&self) -> Result<&PublicKeySet, RoutingError>

Returns the current BLS public key set or RoutingError::InvalidState if we are not joined yet.

pub fn secret_key_share(&self) -> Result<&SecretKeyShare, RoutingError>

Returns the current BLS secret key share or RoutingError::InvalidState if we are not elder.

pub fn our_index(&self) -> Result<usize, RoutingError>

Returns our index in the current BLS group or RoutingError::InvalidState if section key was not generated yet.