Struct routing::Routing

pub struct Routing {
    // some fields omitted
}

Routing provides an actionable interface to RoutingNode. On constructing a new Routing object a RoutingNode will also be started. Routing objects are clonable for multithreading, or a Routing object can be cloned with a new set of keys while preserving a single RoutingNode.

Methods

impl Routing

fn new(event_sender: Sender<Event>) -> Result<Routing, RoutingError>

Starts a new RoutingIdentity, which will also start a new RoutingNode. The RoutingNode will attempt to achieve full routing node status. The intial Routing object will have newly generated keys

fn send_get_request(&self, src: Authority, dst: Authority, content: RequestContent) -> Result<(), InterfaceError>

Send a Get message with a DataRequest to an Authority, signed with given keys.

fn send_put_request(&self, src: Authority, dst: Authority, content: RequestContent) -> Result<(), InterfaceError>

Add something to the network

fn send_post_request(&self, src: Authority, dst: Authority, content: RequestContent) -> Result<(), InterfaceError>

Change something already on the network

fn send_delete_request(&self, src: Authority, dst: Authority, content: RequestContent) -> Result<(), InterfaceError>

Remove something from the network

fn send_get_response(&self, src: Authority, dst: Authority, content: ResponseContent) -> Result<(), InterfaceError>

Respond to a get_request (no error can be sent) If we received the request from a group, we'll not get the signed_request.

fn send_put_response(&self, src: Authority, dst: Authority, content: ResponseContent) -> Result<(), InterfaceError>

response error to a put request

fn send_post_response(&self, src: Authority, dst: Authority, content: ResponseContent) -> Result<(), InterfaceError>

Response error to a post request

fn send_delete_response(&self, src: Authority, dst: Authority, content: ResponseContent) -> Result<(), InterfaceError>

response error to a delete respons

fn send_refresh_request(&self, _type_tag: u64, _src: Authority, _content: Vec<u8>, _cause: XorName) -> Result<(), InterfaceError>

Refresh the content in the close group nodes of group address content::name. This method needs to be called when churn is triggered. all the group members need to call this, otherwise it will not be resolved as a valid content. If the authority provided (src) is not a group, the request for refresh will be dropped.

fn stop(&self)

Signal to RoutingNode that it needs to refuse new messages and handle all outstanding messages. After handling all messages it will send an Event::Terminated to the user.