Struct routing::RoutingClient

pub struct RoutingClient {
    // some fields omitted
}

Routing provides an actionable interface to RoutingNode. On constructing a new Routing object a RoutingNode will also be started. Routing objects are clonable for multithreading, or a Routing object can be cloned with a new set of keys while preserving a single RoutingNode.

Methods

impl RoutingClient

fn new(event_sender: Sender<Event>, keys: Option<FullId>) -> Result<RoutingClient, RoutingError>

Starts a new RoutingIdentity, which will also start a new RoutingNode. The RoutingNode will only bootstrap to the network and not attempt to achieve full routing node status. If the client is started with a relocated id (ie the name has been reassigned), the core will instantely instantiate termination of the client.

fn send_get_request(&mut self, dst: Authority, data_request: DataRequest) -> Result<(), InterfaceError>

Send a Get message with a DataRequest to an Authority, signed with given keys.

fn send_put_request(&self, dst: Authority, data: Data) -> Result<(), InterfaceError>

Add something to the network

fn send_post_request(&self, dst: Authority, data: Data) -> Result<(), InterfaceError>

Change something already on the network

fn send_delete_request(&self, dst: Authority, data: Data) -> Result<(), InterfaceError>

Remove something from the network