Struct routing::StructuredData

pub struct StructuredData {
    // some fields omitted
}

Mutable structured data.

The name is computed from the type tag and identifier, so these two fields are immutable.

These types may be stored unsigned with previous and current owner keys set to the same keys. Updates require a signature to validate.

Methods

impl StructuredData

fn new(type_tag: u64, identifier: XorName, version: u64, data: Vec<u8>, current_owner_keys: Vec<PublicKey>, previous_owner_keys: Vec<PublicKey>, signing_key: Option<&SecretKey>) -> Result<StructuredData, RoutingError>

Creates a new StructuredData signed with signing_key.

fn compute_name(type_tag: u64, identifier: &XorName) -> XorName

This is a static function that computes the name of a StructuredData given its type tag and identifier. To request the data with that type tag and identifier, a Get request needs to be sent to that name's NaeManager.

fn replace_with_other(&mut self, other: StructuredData) -> Result<(), RoutingError>

Replaces this data item with the given updated version if the update is valid, otherwise returns an error.

This allows types to be created and previous_owner_signatures added one by one. To transfer ownership, the current owner signs over the data; the previous owners field must have the previous owners of version - 1 as the current owners of that last version.

fn name(&self) -> XorName

Returns the name, computed from the type tag and identifier.

fn validate_self_against_successor(&self, other: &StructuredData) -> Result<(), RoutingError>

Verifies that other is a valid update for self; returns an error otherwise.

An update is valid if it doesn't change type tag or identifier (these are immutable), increases the version by 1 and is signed by (more than 50% of) the owners.

In case of an ownership transfer, the previous_owner_keys in other must match the current_owner_keys in self.

fn add_signature(&mut self, secret_key: &SecretKey) -> Result<usize, RoutingError>

Adds a signature with the given secret_key to the previous_owner_signatures and returns the number of signatures that are still required. If more than 50% of the previous owners have signed, 0 is returned and validation is complete.

fn replace_signatures(&mut self, new_signatures: Vec<Signature>)

Overwrite any existing signatures with the new signatures provided.

fn get_type_tag(&self) -> u64

Get the type_tag

fn get_identifier(&self) -> &XorName

Get the identifier

fn get_data(&self) -> &Vec<u8>

Get the serialised data

fn get_previous_owner_keys(&self) -> &Vec<PublicKey>

Get the previous owner keys

fn get_version(&self) -> u64

Get the version

fn get_owner_keys(&self) -> &Vec<PublicKey>

Get the current owner keys

fn get_previous_owner_signatures(&self) -> &Vec<Signature>

Get previous owner signatures

fn payload_size(&self) -> usize

Return data size.