Struct routing::structured_data::StructuredData

pub struct StructuredData {
    // some fields omitted
}

StructuredData These types may be stored unsigned with previous and current owner keys set to the same keys. Updates though require a signature to validate

Methods

impl StructuredData

fn new(type_tag: u64, identifier: NameType, version: u64, data: Vec<u8>, current_owner_keys: Vec<PublicKey>, previous_owner_keys: Vec<PublicKey>, signing_key: Option<&SecretKey>) -> Result<StructuredData, RoutingError>

Constructor

fn compute_name(type_tag: u64, identifier: &NameType) -> NameType

This is a static function required for computing a name give the tag-type and identifier to be use by GETs

fn replace_with_other(&mut self, other: StructuredData) -> Result<(), RoutingError>

replace this data item with an updated version if such exists, otherwise fail. This is done this way to allow types to be created and previous_owner_signatures added one by one To transfer ownership the current owner signs over the data, the previous owners field must have the previous owners of version - 1 as the current owners of that last version.

fn name(&self) -> NameType

Returns name and validates invariants

fn validate_self_against_successor(&self, other: &StructuredData) -> Result<(), RoutingError>

Validate...

fn add_signature(&mut self, secret_key: &SecretKey) -> Result<isize, RoutingError>

Returns number of previous_owner_signatures still required (if any, 0 means this is complete)

fn replace_signatures(&mut self, new_signatures: Vec<Signature>)

Overwrite any existing signatures with the new signatures provided

fn get_type_tag(&self) -> u64

Get the type_tag

fn get_identifier(&self) -> &NameType

Get the identifier

fn get_data(&self) -> &Vec<u8>

Get the serialised data

fn get_previous_owner_keys(&self) -> &Vec<PublicKey>

Get the previous owner keys

fn get_version(&self) -> u64

Get the version

fn get_owner_keys(&self) -> &Vec<PublicKey>

Get the current owner keys

fn get_previous_owner_signatures(&self) -> &Vec<Signature>

Get previous owner signatures

fn payload_size(&self) -> usize

Return data size.