Module routing::utils

Commonly required functions.

Functions

calculate_relocated_name
relocated_name = Hash(original_name + 1st closest node id + 2nd closest node id) In case of only one close node provided (in initial network setup scenario), relocated_name = Hash(original_name + 1st closest node id)

decode
Decode a vcetor of bytes to a value of type T, otherwise error on failure.

encode
Encode a value of type T to a vector of bytes.

get_debug_id
Formatted string from a vector of bytes.

public_key_to_client_name
The name client name is the SHA512 of the public signing key