Crate routing

The main API for routing nodes (this is where you give the network its rules)

The network will report from authority your authority and validate cryptographically any message via group consensus. This means any facade you implement will set out what you deem to be a valid operation. Routing will provide a valid message sender and authority that will allow you to set up many decentralised services

The data types are encoded with Concise Binary Object Representation (CBOR).

This allows certain tags to be available to routing, facilitating fields such as data.name(), when calculating authority.

We use Iana tag representations http://www.iana.org/assignments/cbor-tags/cbor-tags.xhtml

Please define your own for this library. These tags are non optional and your data MUST meet the requirements and implement the following tags:

tag: 5483_0 -> name [u8; 64] type
tag: 5483_1 -> XXXXXXXXXXXXXX

Modules

authority

client_interface

error

node_interface

refresh_accumulator

routing_client

routing_membrane
Routing membrane is a single thread responsible for the in- and outgoing messages. It accepts messages received from CRUST. The membrane evaluates whether a message is to be forwarded, or accepted into the membrane as a request where Sentinel holds it until verified and resolved. Requests resolved by Sentinel, will be handed on to the Interface for actioning. A limited number of messages are deliberatly for Routing and network management purposes. Some network management messages are directly handled without Sentinel resolution. Other network management messages are handled by Routing after Sentinel resolution.

routing_node

sendable

test_utils

types

Macros

container_of_u8_to_array!
Convert a container of u8s to an array. If the container is not the exact size specified, None is returned. Otherwise, all of the elements are moved into the array.

Structs

NameType
NameType can be created using the new function by passing ID as it’s parameter.

Functions

closer_to_target
Returns true if lhs is closer to target than rhs. "Closer" here is as per the Kademlia notion of XOR distance, i.e. the distance between two NameTypes is the bitwise XOR of their values.