Crate routing

Client and node implementations for a resilient decentralised network.

The network is based on the kademlia_routing_table and uses the XOR metric to define the "distance" between two XorNames. XorNames are used as addresses of nodes, clients as well as data.

Messages are exchanged between authorities, where an Authority can be an individual client or node, or a group of nodes. In both cases, messages are cryptographically signed by the sender, and in the latter case it is verified that a sufficient number of group members agree on the message: Only if that quorum is reached, the message is delivered. In addition, each message has a unique ID, and is delivered only once.

Group authorities are also addressed using a single XorName. The members of that group are the nodes that are closest to that name. Since nodes are assigned their name by the network, this provides redundancy and resilience: A node has no control over which group authorities it will be a member of, and without a majority in the group it cannot forge a message from that group.

The library also provides different types for the messages' data. The data types are encoded with Concise Binary Object Representation (CBOR).

Usage

A decentralised service based on the routing library uses Client to send requests to the network of nodes and receive responses.

Node is used to handle and send requests within that network, and to implement its functionality, e. g. storing and retrieving data, validating permissions, managing metadata etc.

Client creation

A client's name is a hash of its public keys. Upon creation, the client will attempt to connect to the network through any node, and exchange public keys with it. That node becomes a bootstrap node for the client, and messages to and from the client will be routed over it.

use std::sync::mpsc;
use routing::{Client, Event, FullId};

let (sender, _receiver) = mpsc::channel::<Event>();
let full_id = FullId::new(); // Generate new keys.
let _ = Client::new(sender, Some(full_id.clone())).unwrap();

let _ = full_id.public_id().name();

Messages can be sent using the methods of client, and received as Events from the receiver.

Node creation

Creating a node looks even simpler:

use std::sync::mpsc;
use routing::{Node, Event};

let (sender, _receiver) = mpsc::channel::<Event>();
let _ = Node::new(sender).unwrap();

Upon creation, the node will first connect to the network as a client. Once it has client status, it requests a new name from the network, and then integrates itself in the network with that new name, adding close nodes to its routing table.

Messages can be sent using the methods of node, and received as Events from the receiver. The node can act as an individual node or as part of a group authority. Sending a message as a group authority only has an effect if sufficiently many other nodes in that authority send the same message.

Example: Put message flow for immutable data

An example implementation of a vault for storing and retrieving ImmutableData is included with the code of this library. The flow of a Put request there moves through the following authorities:

A Client that wants to store a piece of data on the network.

The ClientManager group authority that is responsible for that particular client. The nodes in this group have access to metadata about this client and can e. g. verify that it has required permissions or quota to store the data. The ClientManager's address is the XorName of the client itself, so it consists of the nodes closest to that name.

The NaeManager group authority which is responsible for that particular piece of data and knows where in the network it should be stored. Its address is the XorName of the data.

The ManagedNodes are the individual nodes that actually store the data. This could be just one, but let's assume that data is stored redundantly in more than one place.

The message flow from the client to the managed nodes looks like this:

The client sends a Put request to the ClientManager, i. e. the sender is the individual Client and the receiver is a group of nodes with addresses close to the client's name.

Each node in the ClientManager group performs validation and updates metadata (which may involve further messages not considered in this example). If valid, it sends a PutSuccess response back to the client, and a Put request to the NaeManager with the data's XorName. In both cases, the sender is the ClientManager, not the individual node.

Each node in the NaeManager chooses a number of individual nodes that are responsible for storing this specific piece of data. This needs to happen in a deterministic way so that the members of the NaeManager agree on the responsible nodes. Then they finally send a Put request to each of these. Here the sender is the NaeManager, but the recipient is the individual ManagedNode.

Each ManagedNode locally stores the data.

These steps need to be implemented by the user of the library. Each node needs to handle all these cases, as it can be simultaneously a ManagedNode and a member of several different ClientManagers and NaeManagers. A client, however, does not act as a node or vice versa.

Transparently to the user, the routing library deals with message quora, so the user will only receive a message from a Client or Node if it was either sent

by a client or individual node, or

by a sufficient number of members of a group authority.

Structs

Client
Interface for sending and receiving messages to and from a network of nodes in the role of a client.

FullId
Network identity component containing name, and public and private keys.

ImmutableData
An immutable chunk of data.

MessageId
Unique ID for messages

Node
Interface for sending and receiving messages to and from other nodes, in the role of a full routing node.

PlainData
Plain data with a name and a value.

PublicId
Network identity component containing name and public keys.

RequestMessage
A request message wrapper

ResponseMessage
A response message wrapper

SignedMessage
Wrapper around a routing message, signed by the originator of the message.

StructuredData
Mutable structured data.

Enums

Authority
An entity that can act as a source or destination of a message.

Data
This is the data types routing handles in the public interface

DataRequest
DataRequest.

Event
An Event raised by a Node or Client via its event sender.

ImmutableDataType
The type of an individual copy of immutable data.

InterfaceError
InterfaceError.

RequestContent
The request types

ResponseContent
The response types

RoutingError
RoutingError.

RoutingMessage
Variant type to hold either a request or response.

Constants

MAX_STRUCTURED_DATA_SIZE_IN_BYTES
Maximum allowed size for a Structured Data to grow to