Crate routing

The API for routing nodes (this is where you give the network its rules)

The network will report from authority your authority and validate cryptographically any message via group consensus or direct (clients). This means any facade you implement will set out what you deem to be a valid operation. Routing will provide

Valid message sender

Confirmed from authority

Confirmed your authhority

Exaclty 1 copy of each message

This should allow relatively easy set up many decentralised services. Setting rules for data types and what can be done wiht such types at differnt personas will allow a fairly complex network to be configured wiht relative ease.

The data types are encoded with Concise Binary Object Representation (CBOR).

Modules

test_utils
TODO Remove this from public visibility

types
Types and functions used throughout the library.

Structs

FullId
Network identity component containing name, and public and private keys.

ImmutableData
ImmutableData hash == SHA512

PlainData
PlainData

PublicId
PublicId.

RequestMessage
A request message wrapper

ResponseMessage
A response message wrapper

Routing
Routing provides an actionable interface to RoutingNode. On constructing a new Routing object a RoutingNode will also be started. Routing objects are clonable for multithreading, or a Routing object can be cloned with a new set of keys while preserving a single RoutingNode.

RoutingClient
Routing provides an actionable interface to RoutingNode. On constructing a new Routing object a RoutingNode will also be started. Routing objects are clonable for multithreading, or a Routing object can be cloned with a new set of keys while preserving a single RoutingNode.

SignedMessage
Wrapper around a routing message, signed by the originator of the message.

StructuredData
StructuredData These types may be stored unsigned with previous and current owner keys set to the same keys. Updates though require a signature to validate

Enums

Authority
Persona types recognised by network.

Data
This is the data types routing handles in the public interface

DataRequest
DataRequest.

Event
An Event is received at the effective close group of B of a message flow < A | B >

ImmutableDataType
ImmutableDataType.

InterfaceError
InterfaceError.

RequestContent
types of requests allowed on network

ResponseContent
Types of respnses to exepect on the network. All responses will map to a specific request and where request was from a single node or client the response will contatin the singed request. This prevents forgery or co-ersion attacks.

RoutingError
RoutingError.

RoutingMessage
Variant type to old either a request or response

Constants

MAX_STRUCTURED_DATA_SIZE_IN_BYTES
Maximum allowed size for a Structured Data to grow to