Crate routing

The main API for routing nodes (this is where you give the network its rules)

The network will report from authority your authority and validate cryptographically any message via group consensus. This means any facade you implement will set out what you deem to be a valid operation. Routing will provide a valid message sender and authority that will allow you to set up many decentralised services.

The data types are encoded with Concise Binary Object Representation (CBOR).

We use Iana tag representations http://www.iana.org/assignments/cbor-tags/cbor-tags.xhtml

Reexports

pub use authority::Authority;

Modules

authority
Persona types recognised by network.

data
Data types used in messages.

error
Errors reported for failed conditions/operations.

event
Event provides the events the user can expect to receive from routing.

id
Private network identity component.

immutable_data
ImmutableData type.

plain_data
PlainData type.

public_id
Public network identity component.

routing
Routing provides an actionable interface to routing.

routing_client
Client interface to routing.

structured_data
StructuredData type.

test_utils
Utility structs and functions used during testing.

types
Types and functions used throught the library.

utils
Commonly required functions.

Structs

NameType
NameType can be created using the new function by passing ID as it’s parameter.

SignedToken
SignedToken.

Enums

ExternalRequest
These are the message types routing provides. ExternalRequest.

ExternalResponse
ExternalResponse.

Constants

NAME_TYPE_LEN
Constant byte length of NameType.

Functions

closer_to_target
Returns true if lhs is closer to target than rhs. "Closer" here is as per the Kademlia notion of XOR distance, i.e. the distance between two NameTypes is the bitwise XOR of their values.