Enum routing::AppendWrapper

pub enum AppendWrapper {
    Pub {
        append_to: XorName,
        data: AppendedData,
        version: u64,
    },
    Priv {
        append_to: XorName,
        data: PrivAppendedData,
        sign_key: PublicKey,
        version: u64,
        signature: Signature,
    },
}

An AppendedData item, together with the identifier of the data to append it to.

Variants
Pub

A wrapper for public appendable data.

Fields

append_to: XorName

The name of the data chunk to add to.

data: AppendedData

The item to add to the chunk.

version: u64

The current version of the chunk.

Priv

A wrapper for private appendable data.

Fields

append_to: XorName

The name of the data chunk to add to.

data: PrivAppendedData

The item to add to the chunk.

sign_key: PublicKey

The signer's public_key.

version: u64

The current version of the chunk.

signature: Signature

All the above fields

Methods

impl AppendWrapper

fn new_pub(append_to: XorName, data: AppendedData, version: u64) -> Self

Returns a new append wrapper for public data.

fn new_priv(append_to: XorName, data: PrivAppendedData, sign_pair: (&PublicKey, &SecretKey), version: u64) -> Result<AppendWrapper, RoutingError>

Returns a new, signed append wrapper for private data.

fn identifier(&self) -> DataIdentifier

Returns the identifier of the data to append to.

fn verify_signature(&self) -> bool

Returns true if the signature matches the data.

fn sign_key(&self) -> &PublicKey

Returns sign_key of the signer.

fn version(&self) -> &u64

Returns version of the wrapper item.

fn priv_appended_data(&self) -> Option<&PrivAppendedData>

Returns priv_appended_data if AppendWrapper::Priv.

fn pub_appended_data(&self) -> Option<&AppendedData>

Returns pub_appended_data if AppendWrapper::Pub.