Enum routing::Authority

pub enum Authority {
    ClientManager(XorName),
    NaeManager(XorName),
    NodeManager(XorName),
    ManagedNode(XorName),
    Client {
        client_key: PublicKey,
        proxy_node_name: XorName,
    },
}

An entity that can act as a source or destination of a message.

An Authority can be an individual Client or ManagedNode, or a group of nodes, like a NodeManager, ClientManager or NaeManager.

Variants

ClientManager

Manager of a Client. XorName is the hash of the Client's client_key.

NaeManager

Manager of a network-addressable element. XorName is the name of the element in question.

NodeManager

Manager of a ManagedNode. XorName is that of the ManagedNode.

ManagedNode

A non-client node (i.e. a vault) which is managed by NodeManagers. XorName is provided by the network relocation process immediately after bootstrapping.

Client

A Client.

Fields

client_key

The client's public signing key. The hash of this specifies the location of the Client in the network address space.

proxy_node_name

The name of the single ManagedNode which the Client connects to and proxies all messages through.

Methods

impl Authority

fn is_group(&self) -> bool

Returns true if group authority, otherwise false.

fn get_name(&self) -> &XorName

Returns the name of authority.