Enum routing::Event

pub enum Event {
    Request(RequestMessage),
    Response(ResponseMessage),
    Churn {
        id: MessageId,
        lost_close_node: Option<XorName>,
    },
    Connected,
}

An Event raised by a Node or Client via its event sender.

These are sent by routing to the library's user. It allows the user to handle requests and responses, and to react to changes in the network.

Request and Response events from group authorities are only raised once the quorum has been reached, i. e. enough members of the group have sent the same message.

Variants

Request

Request.

Response

Response.

Churn

A churn event: a node left or joined this node's close group.

Fields

id

The unique ID of this Churn event.

lost_close_node

The name of the node that left the close group, if any.

Connected

The client has successfully connected to a proxy node on the network.