Enum routing::RequestContent

pub enum RequestContent {
    GetNetworkName {
        current_id: PublicId,
    },
    ExpectCloseNode {
        expect_id: PublicId,
    },
    GetCloseGroup,
    Connect,
    Endpoints {
        encrypted_endpoints: Vec<u8>,
        nonce_bytes: [u8; box_::NONCEBYTES],
    },
    GetPublicId,
    GetPublicIdWithEndpoints {
        encrypted_endpoints: Vec<u8>,
        nonce_bytes: [u8; box_::NONCEBYTES],
    },
    Refresh {
        type_tag: u64,
        message: Vec<u8>,
        cause: XorName,
    },
    Get(DataRequest),
    Put(Data),
    Post(Data),
    Delete(Data),
}

types of requests allowed on network

Variants

GetNetworkName

Ask network to alter your PublciId name and forward to appropriate group Client -> NaeManager

Fields

current_id

Your PublicId (public keys and name)

ExpectCloseNode

Receiving group of relocatted name will get this message

Fields

expect_id

Your PublicId (public keys and name)

GetCloseGroup

Ask each memeber of a group near an address for the PublicId of that node

Connect

Request a connection to this node

Endpoints

Send our endpoints encrypted toa node we wish to connect to and have the keys for

Fields

encrypted_endpoints

encrypted crust endpoints (socket addr and protocol)

nonce_bytes

nonce used to provide a salt in the encrytped message

GetPublicId

Ask each memeber of a group near a node address for the PublicId

GetPublicIdWithEndpoints

Ask for a publicId but provide our endpoints encrytped

Fields

encrypted_endpoints

encrypted crust endpoints (socket addr and protocol)

nonce_bytes

nonce used to provide a salt in the encrytped message

Refresh

Message from upper layers sending network state on any network churn event

Fields

type_tag

externally defined type identifier

message

externally defined message

cause

The node that caused the churn event. Used here (passed up to upper layers in churn event) who must give it back in which allows filtering of different churn events (used as unique identifier)

Get

Ask for data from network, passed from API with data name as parameter

Put

Put data to network. Provide actual data as parameter

Post

Post data to network. Provide actual data as parameter

Delete

Delete data from network. Provide actual data as parameter