Enum routing::RoutingError

pub enum RoutingError {
    NotBootstrapped,
    Terminated,
    BadAuthority,
    AlreadyConnected,
    UnknownMessageType,
    FailedSignature,
    NotEnoughSignatures,
    DuplicateSignatures,
    FilterCheckFailed,
    FailedToBootstrap,
    RoutingTableEmpty,
    RejectedPublicId,
    RefusedFromRoutingTable,
    RefreshNotFromGroup,
    Utf8(Utf8Error),
    Interface(InterfaceError),
    Io(Error),
    Cbor(CborError),
    InvalidStateForOperation,
    SerialisationError(SerialisationError),
    AsymmetricDecryptionFailure,
    UnknownConnection,
    DirectionCheckFailed,
    RoutingTableBucketIndexFailed,
    InvalidDestination,
    ProxyConnectionNotFound,
    ClientConnectionNotFound,
    InvalidSource,
}

RoutingError.

Variants

NotBootstrapped

The node/client has not bootstrapped yet

Terminated

Invalid State

BadAuthority

invalid requester or handler authorities

AlreadyConnected

failure to connect to an already connected node

UnknownMessageType

received message having unknown type

FailedSignature

Failed signature check

NotEnoughSignatures

Not Enough signatures

DuplicateSignatures

Duplicate signatures

FilterCheckFailed

duplicate request received

FailedToBootstrap

failure to bootstrap off the provided endpoints

RoutingTableEmpty

unexpected empty routing table

RejectedPublicId

public id rejected because of unallowed relocated status

RefusedFromRoutingTable

routing table did not add the node information, either because it was already added, or because it did not improve the routing table

RefreshNotFromGroup

We received a refresh message but it did not contain group source address

Utf8

String errors

Interface

interface error

Io

i/o error

Cbor

serialisation error

InvalidStateForOperation

Current state is invalid for the operation

SerialisationError

Serialisation Error

AsymmetricDecryptionFailure

Asymmetric Decryption Failure

UnknownConnection

Unknown Connection

DirectionCheckFailed

The message is not getting closer to the target

RoutingTableBucketIndexFailed

Density mismatch

InvalidDestination

Invalid Destination

ProxyConnectionNotFound

Connection to proxy node does not exist in proxy map

ClientConnectionNotFound

Connection to client does not exist in client map

InvalidSource

Invalid Source