Enum routing::RoutingError

pub enum RoutingError {
    NotBootstrapped,
    Terminated,
    BadAuthority,
    AlreadyConnected,
    UnknownMessageType,
    FailedSignature,
    NotEnoughSignatures,
    DuplicateSignatures,
    InvalidOwners,
    FilterCheckFailed,
    FailedToBootstrap,
    RejectedPublicId,
    RejectedClientMessage,
    RoutingTable(RoutingTableError),
    Utf8(Utf8Error),
    Interface(InterfaceError),
    Io(Error),
    Crust(CrustError),
    SendEventError(SendError<Event>),
    InvalidStateForOperation,
    SerialisationError(SerialisationError),
    AsymmetricDecryptionFailure,
    UnknownConnection(PeerId),
    InvalidDestination,
    ProxyConnectionNotFound,
    ClientConnectionNotFound,
    InvalidSource,
    CannotTunnelThroughTunnel,
    HashMismatch,
}

The type of errors that can occur during handling of routing events.

Variants
NotBootstrapped

The node/client has not bootstrapped yet

Terminated

Invalid State

BadAuthority

Invalid requester or handler authorities

AlreadyConnected

Failure to connect to an already connected node

UnknownMessageType

Received message having unknown type

FailedSignature

Failed signature check

NotEnoughSignatures

Not Enough signatures

DuplicateSignatures

Duplicate signatures

InvalidOwners

The list of owner keys is invalid

FilterCheckFailed

Duplicate request received

FailedToBootstrap

Failure to bootstrap off the provided endpoints

RejectedPublicId

Public id rejected because of disallowed relocated status

RejectedClientMessage

A client with client_restriction == true tried to send a message restricted to nodes.

RoutingTable(RoutingTableError)

Routing Table error

Utf8(Utf8Error)

String errors

Interface(InterfaceError)

Interface error

Io(Error)

i/o error

Crust(CrustError)

Crust error

SendEventError(SendError<Event>)

Channel sending error

InvalidStateForOperation

Current state is invalid for the operation

SerialisationError(SerialisationError)

Serialisation Error

AsymmetricDecryptionFailure

Asymmetric Decryption Failure

UnknownConnection(PeerId)

Unknown Connection

InvalidDestination

Invalid Destination

ProxyConnectionNotFound

Connection to proxy node does not exist in proxy map

ClientConnectionNotFound

Connection to client does not exist in client map

InvalidSource

Invalid Source

CannotTunnelThroughTunnel

Attempted to use a node as a tunnel that is not directly connected

HashMismatch

Decoded a user message with an unexpected hash.