Enum routing::event::Event

pub enum Event {
    Request {
        request: ExternalRequest,
        our_authority: Authority,
        from_authority: Authority,
        response_token: Option<SignedToken>,
    },
    Response {
        response: ExternalResponse,
        our_authority: Authority,
        from_authority: Authority,
    },
    FailedRequest {
        request: ExternalRequest,
        our_authority: Option<Authority>,
        location: Authority,
        interface_error: InterfaceError,
    },
    FailedResponse {
        response: ExternalResponse,
        our_authority: Option<Authority>,
        location: Authority,
        interface_error: InterfaceError,
    },
    Refresh(u64, NameType, Vec<Vec<u8>>),
    Churn(Vec<NameType>),
    Bootstrapped,
    Connected,
    Disconnected,
    Terminated,
}

An Event is received at the effective close group of B of a message flow < A | B > 1. Event::MessageSecured provides the RoutingMessage after being secured by routing, our authority provides the base authority as validated by routing. When the from_authority in routing message is a Client or ManagedNode, the original signature is provided for reply or response. 2. Event::Refresh has accumulated Refresh messages centered on a name for a type_tag. This can be used to transfer accounts between nodes of an effective close group. 3. Event::Churn occurs when our close group changes. The new close group is provided. Our close group always contains our own name first. When we are connected to other nodes the list contains minimally two names. 4. Event::Terminated is called after RoutingNode::stop() has ensured all message queues are processed and empty.

Variants

Request
Fields

request

our_authority

from_authority

response_token

Response
Fields

response

our_authority

from_authority

FailedRequest
Fields

request

our_authority

location

interface_error

FailedResponse
Fields

response

our_authority

location

interface_error

Refresh

Churn

Bootstrapped

Connected

Disconnected

Terminated