Struct routing::routing_client::RoutingClient

pub struct RoutingClient {
    // some fields omitted
}

Routing provides an actionable interface to RoutingNode. On constructing a new Routing object a RoutingNode will also be started. Routing objects are clonable for multithreading, or a Routing object can be cloned with a new set of keys while preserving a single RoutingNode.

Methods

impl RoutingClient

fn new(event_sender: Sender<Event>, keys: Option<Id>) -> RoutingClient

Starts a new RoutingIdentity, which will also start a new RoutingNode. The RoutingNode will only bootstrap to the network and not attempt to achieve full routing node status. If the client is started with a relocated id (ie the name has been reassigned), the core will instantely instantiate termination of the client.

fn get_request(&mut self, location: Authority, data_request: DataRequest)

Send a Get message with a DataRequest to an Authority, signed with given keys.

fn put_request(&self, location: Authority, data: Data)

Add something to the network

fn post_request(&self, location: Authority, data: Data)

Change something already on the network

fn delete_request(&self, location: Authority, data: Data)

Remove something from the network

fn stop(&mut self)

Signal to RoutingNode that it needs to refuse new messages and handle all outstanding messages. After handling all messages it will send an Event::Terminated to the user.