Struct routing::Node

pub struct Node {
    // some fields omitted
}

Node provides an actionable interface to Core. On constructing a new Node object a Core will also be started.

Methods

impl Node

fn new(event_sender: Sender<Event>) -> Result<Node, RoutingError>

Starts a new RoutingIdentity, which will also start a new Core. The Core will attempt to achieve full routing node status. The intial Node object will have newly generated keys

fn send_get_request(&self, src: Authority, dst: Authority, data_request: DataRequest, id: MessageId) -> Result<(), InterfaceError>

Send a Get message with a DataRequest to an Authority, signed with given keys.

fn send_put_request(&self, src: Authority, dst: Authority, data: Data, id: MessageId) -> Result<(), InterfaceError>

Add something to the network

fn send_post_request(&self, src: Authority, dst: Authority, data: Data, id: MessageId) -> Result<(), InterfaceError>

Change something already on the network

fn send_delete_request(&self, src: Authority, dst: Authority, data: Data, id: MessageId) -> Result<(), InterfaceError>

Remove something from the network

fn send_get_success(&self, src: Authority, dst: Authority, data: Data, id: MessageId) -> Result<(), InterfaceError>

Respond to a get_request indicating success

fn send_get_failure(&self, src: Authority, dst: Authority, request: RequestMessage, external_error_indicator: Vec<u8>, id: MessageId) -> Result<(), InterfaceError>

Respond to a get_request indicating failure

fn send_put_success(&self, src: Authority, dst: Authority, request_hash: Digest, id: MessageId) -> Result<(), InterfaceError>

Respond to a put_request indicating success

fn send_put_failure(&self, src: Authority, dst: Authority, request: RequestMessage, external_error_indicator: Vec<u8>, id: MessageId) -> Result<(), InterfaceError>

Respond to a put_request indicating failure

fn send_post_success(&self, src: Authority, dst: Authority, request_hash: Digest, id: MessageId) -> Result<(), InterfaceError>

Respond to a post_request indicating success

fn send_post_failure(&self, src: Authority, dst: Authority, request: RequestMessage, external_error_indicator: Vec<u8>, id: MessageId) -> Result<(), InterfaceError>

Respond to a post_request indicating failure

fn send_delete_success(&self, src: Authority, dst: Authority, request_hash: Digest, id: MessageId) -> Result<(), InterfaceError>

Respond to a delete_request indicating success

fn send_delete_failure(&self, src: Authority, dst: Authority, request: RequestMessage, external_error_indicator: Vec<u8>, id: MessageId) -> Result<(), InterfaceError>

Respond to a delete_request indicating failure

fn send_refresh_request(&self, src: Authority, content: Vec<u8>) -> Result<(), InterfaceError>

Refresh the content in the close group nodes of group address in src. This method needs to be called when churn is triggered. all the group members need to call this, otherwise it will not be resolved as a valid content. The authority provided (src) should be a group.

fn close_group(&self) -> Result<Vec<XorName>, InterfaceError>

Returns the names of the close group to this node.

fn name(&self) -> Result<XorName, InterfaceError>

Returns the name of this node.