Struct routing::Node

pub struct Node {
    // some fields omitted
}

Interface for sending and receiving messages to and from other nodes, in the role of a full routing node.

A node is a part of the network that can route messages and be member of a group authority. Its methods can be used to send requests and responses as either an individual ManagedNode or as a part of a group authority. Their src argument indicates that role, so it must always either be the ManagedNode with this node's name, or the ClientManager or NodeManager or NaeManager with the address of a client, node or data element that this node is close to.

Methods

impl Node

fn new(event_sender: Sender<Event>) -> Result<Node, RoutingError>

Create a new Node.

It will automatically connect to the network in the same way a client does, but then request a new name and integrate itself into the network using the new name.

The intial Node object will have newly generated keys.

fn send_get_request(&self, src: Authority, dst: Authority, data_request: DataRequest, id: MessageId) -> Result<(), InterfaceError>

Send a Get request to dst to retrieve data from the network.

fn send_put_request(&self, src: Authority, dst: Authority, data: Data, id: MessageId) -> Result<(), InterfaceError>

Send a Put request to dst to store data on the network.

fn send_post_request(&self, src: Authority, dst: Authority, data: Data, id: MessageId) -> Result<(), InterfaceError>

Send a Post request to dst to modify data on the network.

fn send_delete_request(&self, src: Authority, dst: Authority, data: Data, id: MessageId) -> Result<(), InterfaceError>

Send a Delete request to dst to remove data from the network.

fn send_get_success(&self, src: Authority, dst: Authority, data: Data, id: MessageId) -> Result<(), InterfaceError>

Respond to a Get request indicating success and sending the requested data.

fn send_get_failure(&self, src: Authority, dst: Authority, request: RequestMessage, external_error_indicator: Vec<u8>, id: MessageId) -> Result<(), InterfaceError>

Respond to a Get request indicating failure.

fn send_put_success(&self, src: Authority, dst: Authority, request_hash: Digest, id: MessageId) -> Result<(), InterfaceError>

Respond to a Put request indicating success.

fn send_put_failure(&self, src: Authority, dst: Authority, request: RequestMessage, external_error_indicator: Vec<u8>, id: MessageId) -> Result<(), InterfaceError>

Respond to a Put request indicating failure.

fn send_post_success(&self, src: Authority, dst: Authority, request_hash: Digest, id: MessageId) -> Result<(), InterfaceError>

Respond to a Post request indicating success.

fn send_post_failure(&self, src: Authority, dst: Authority, request: RequestMessage, external_error_indicator: Vec<u8>, id: MessageId) -> Result<(), InterfaceError>

Respond to a Post request indicating failure.

fn send_delete_success(&self, src: Authority, dst: Authority, request_hash: Digest, id: MessageId) -> Result<(), InterfaceError>

Respond to a Delete request indicating success.

fn send_delete_failure(&self, src: Authority, dst: Authority, request: RequestMessage, external_error_indicator: Vec<u8>, id: MessageId) -> Result<(), InterfaceError>

Respond to a Delete request indicating failure.

fn send_refresh_request(&self, src: Authority, content: Vec<u8>) -> Result<(), InterfaceError>

Send a Refresh request from src to src to trigger churn.

This is intended to be sent from a group authority (src) to itself whenever a node joins or leaves the group. If the quorum is reached, i. e. enough members agree that a change has happened, the churn mechanism is triggered to adapt to the change.

fn close_group(&self) -> Result<Vec<XorName>, InterfaceError>

Returns the names of the nodes in the routing table which are closest to this one.

fn name(&self) -> Result<XorName, InterfaceError>

Returns the name of this node.