Struct routing::XorName

pub struct XorName(pub [u8; 32]);

A XOR_NAME_BITS-bit number, viewed as a point in XOR space.

This wraps an array of XOR_NAME_LEN bytes, i. e. a number between 0 and 2XOR_NAME_BITS - 1.

XOR space is the space of these numbers, with the XOR metric as a notion of distance, i. e. the points with IDs x and y are considered to have distance x xor y.

Methods

impl XorName

pub fn to_hex(&self) -> String

Hex-encode the XorName as a String.

pub fn count_differing_bits(&self, other: &XorName) -> u32

Returns the number of bits in which self differs from other.

pub fn from_hex(s: &str) -> Result<XorName, XorNameFromHexError>

Hex-decode a XorName from a &str.

pub fn closer(&self, lhs: &XorName, rhs: &XorName) -> bool

Returns true if lhs is closer to self than rhs.

Equivalently, this returns true if in the most significant bit where lhs and rhs disagree, lhs agrees with self.

pub fn closer_or_equal(&self, lhs: &XorName, rhs: &XorName) -> bool

Returns true if lhs is closer to self than rhs, or lhs == rhs.